% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManovaModel.R
\docType{data}
\name{manova_model}
\alias{manova_model}
\title{Determines the relationship between the hinfoot_length and weight variables by sex concurrently}
\format{
An object of class \code{manova} (inherits from \code{maov}, \code{aov}, \code{mlm}, \code{lm}) of length 13.
}
\usage{
manova_model
}
\arguments{
\item{sex}{Sex of rodents}

\item{hindfoot_length}{Hindfoot Length of rodents}

\item{weight}{Weight of rodents}
}
\value{
Returns the relationship of hindfoot length and weight as statistical values that determine what effect the independent variable has on the dependent variable
}
\description{
Use hindfoot length and weight column from surveys dataset
Returns hindfoot length and weight as statistical values together from rodents surveys dataset
}
\keyword{datasets}
