% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManovaModel.R
\name{manova_project}
\alias{manova_project}
\title{Determines the relationship between the hinfoot_length and weight variables by sex concurrently}
\usage{
manova_project(surveys, weight, hindfoot_length, sex)
}
\arguments{
\item{weight}{Weight of rodents}

\item{hindfoot_length}{Hindfoot Length of rodents}

\item{sex}{Sex of rodents}
}
\value{
Returns the relationship of hindfoot length and weight as statistical values that determine what effect the independent variable has on the dependent variable
}
\description{
Use hindfoot length and weight column from surveys dataset
Returns hindfoot length and weight as statistical values together from rodents surveys dataset
}
