% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCount.R
\name{plotcount_project}
\alias{plotcount_project}
\title{Calculate the number of rodents per plot type}
\usage{
plotcount_project(surveys, plot_type)
}
\arguments{
\item{plot_type}{Indicates which plot used by rodents}
}
\value{
Returns number of rodents found at each plot type
}
\description{
Use plot type column from dataset
Returns number of rodents present at each plot
}
