% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNAs.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove NAs from hindfoot length column}
\usage{
remove_nas(surveys, hindfoot_length)
}
\arguments{
\item{hindfoot_length}{hindfoot length of rodents}
}
\value{
returns hindfoot length of rodents in survey dataset without NAs
}
\description{
Use hindfoot length column from surveys dataset
Returns hindfoot length data complete without NAs
}
