% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDev.R
\docType{data}
\name{weight_sd}
\alias{weight_sd}
\title{Calculates and summarizes the standard deviation and mean of weight by species id}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 3 columns.
}
\usage{
weight_sd
}
\arguments{
\item{species_id}{Identification of rodents by species type}

\item{weight}{Weight of rodents}
}
\value{
Returns the summary of the weight column through the standard deviation and mean by species id
}
\description{
Use weight column from surveys dataset and group by species id
Returns the standard deviation and mean of the weight column by species id
}
\keyword{datasets}
